Fix import references after changing app name (from proversity to custom_reports) and api folder name (from reports to api). Move models from api to app and move migration files as well. Rename api_views to views.
