+++ common/djangoapps/custom_reports/__init__.py
@@ -1,3 +1,3 @@
 """
-init.py file for proversity
+init.py file for custom_reports
 """
